Extract last refreshed time for Power BI datasets
