+++ tests/power_bi/expected.json
@@ -39,7 +39,8 @@
       "workspace": {
         "id": "workspace-1",
         "name": "Workspace"
-      }
+      },
+      "lastRefreshed": "2022-01-01T01:02:03+00:00"
     }
   },
   {
